mcp_server/requirements.txt
mcp[cli]
httpx
faiss-cpu
langchain-community
python-dotenv
pypdf
cloudpathlib[s3,gs,azure]
langchain-aws